Multiple lines
made E.row an array of erow structs.
Move the code in editorOpen that initializes E.row to a new function called editorAppendRow, renamed the line and linelen variables to s and len.
Tell realloc() how many bytes we want to allocate, so multiply the number of bytes each erow takes (sizeof(erow)) and multiply that by the number of rows we want.
Update editorDrawRows() to use E.row[y] instead of E.row, when printing out the current line.
Add a while loop to editorOpen() to read an entire file into E.row.