Vertical Scrolling
Add a rowoff variable to the global editor state. Initialize it to 0.
editorDrawRows() display the correct range of lines of the file according to the value of rowoff.
To get the row of the file that we want to display at each y position, add E.rowoff to the y position.
Add editorScroll() function with the following logic: If the cursor has moved outside of the visible window,and if so, adjust E.rowoff so that the cursor is just inside the visible window.