Rendering Tabs
adding render and rsize to the erow struct, and initializing them in editorAppendRow().
make an editorUpdateRow() function that uses the chars string of an erow to fill in the contents of the render string.
replace chars and size with render and rsize in editorDrawRows().
Add code to editorUpdateRow() that renders tabs as multiple space characters
Make the length of a tab stop a constant.