Tabs and the cursor
Adding rx to the global state strcut, and initializing it to 0.
Set the value of E.rx at the top of editorScroll().
Change E.cx to E.rx in editorRefreshScreen() where we set the cursor position.
Create an editorRowCxToRx() function that converts a chars index into a render index, where for each character,  if it’s a tab we use rx % KILO_TAB_STOP to find out how many columns we are to the right of the last tab stop, and then subtract that from KILO_TAB_STOP - 1 to find out how many columns we are to the left of the next tab stop. We add that amount to rx to get just to the left of the next tab stop, and then the unconditional rx++ statement gets us right on the next tab stop.
Call editorRowCxToRx() at the top of editorScroll() to finally set E.rx to its proper value.