Status bar
Decrement E.screenrows so that editorDrawRows() doesn’t try to draw a line of text at the bottom of the screen. Also have editorDrawRows() print a newline after the last row it draws, since the status bar is now the final line being drawn on the screen.
Add a filename string to the global editor state.
Initialize E.filename to the NULL pointer, and it will stay NULL if a file isn't opened.
Display up to 20 characters of the filename, followed by the number of lines in the file. If there is no filename, we'll display [No name] instead
And show the current line number, and align it to the right edge of the screen.