Save to disk
Add a function that converts our array of erow structs into a single string.
Add up the lengths of each row of text, adding 1 to each one for the newline character we'll add to the end of each line.
Return buf, expecting the caller to free() the memory.
Implement the editorSave() function, which will actually wrtie the string returned by editorRowsToString() to disk.
Add error handling to editorSave().