Search
use editorPrompt() to implement a minimal search feature.
Add editorFind().
Create an editorRowRxToCx() function, which is the opposite of the editorRowCxToRx() function.
Call editorRowRxToCx() to convert the matched index to a chars index and assign that to E.cx
map Ctrl-F to the editorFind() function, and add it to the help message we set in main().