Incremental search
In editorPrompt() take a callback function as an argument.
The if statements allow the caller to pass NULL for the callback, in case they don’t want to use a callback. This is the case when we prompt the user for a filename, so let’s pass NULL to editorPrompt() when we do that.
move the actual searching code from editorFind() into a function called editorFindCallback().